+++ src/graphics/bresenham2d.cpp
@@ -15,8 +15,8 @@
 
 void bresenham2d(int x0, int y0, int x1, int y1, _STD_FUNCTION(void(int, int)) passCommand) {
     
-    int dx = fabs(x1-x0), sx = x0<x1 ? 1 : -1;
-    int dy = fabs(y1-y0), sy = y0<y1 ? 1 : -1;
+    int dx = (int) fabs(x1-x0), sx = x0<x1 ? 1 : -1;
+    int dy = (int) fabs(y1-y0), sy = y0<y1 ? 1 : -1;
     int err = (dx>dy ? dx : -dy)/2, e2;
     
     for(;;){
